+++ src/main/java/com/ssg/meowwms/service/stock/StockService.java
@@ -1,4 +1,4 @@
-package com.ssg.meowwms.service;
+package com.ssg.meowwms.service.stock;
 
 import com.ssg.meowwms.dto.stock.ProductStatusDTO;
 import com.ssg.meowwms.dto.stock.StockDTO;
