+++ src/test/java/com/ssg/meowwms/service/StockServiceTests.java
@@ -1,5 +1,6 @@
 package com.ssg.meowwms.service;
 
+import com.ssg.meowwms.service.stock.StockService;
 import lombok.extern.log4j.Log4j2;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
